requirements.txt

# Japan Fire Dashboard — Python Dependencies

# Web framework + async server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9     # for file uploads (satellite data)

# Database
sqlalchemy>=2.0.30
asyncpg>=0.29.0              # async PostgreSQL driver (FastAPI)
psycopg2-binary>=2.9.9      # sync PostgreSQL driver (ingestion scripts)
geoalchemy2>=0.15.1          # SQLAlchemy types for PostGIS geometries

# Geospatial
geopandas>=0.14.4
shapely==2.1.2
pyproj>=3.6.1               # CRS transformations
rasterio>=1.3.10            # raster I/O (Sentinel-2, DEM)
fiona>=1.9.6                # vector I/O
rtree==1.2.0                # spatial indexing

# Data processing
pandas>=2.2.2
numpy>=1.26.4
scipy>=1.13.0               # spatial statistics

# HTTP / API clients
httpx>=0.27.0               # async HTTP (FIRMS, JMA, Sentinel Hub)
requests>=2.32.2            # sync HTTP (setup scripts)
aiohttp>=3.9.5              # async streaming downloads

# Scheduling (runs inside FastAPI — no Airflow needed)
apscheduler>=3.10.4

# Machine learning (Phase 4)
scikit-learn>=1.5.0
xgboost>=2.0.3
shap>=0.45.0                # SHAP values for model explainability
joblib>=1.4.2               # model serialisation
#mlflow>=2.13.0              # experiment tracking 

# Fire Weather Index calculation
#cffdrs             # Canadian FWI system (pip install cffdrs)

# Configuration
python-dotenv>=1.0.1        # load .env files
pydantic>=2.7.1             # data validation (FastAPI models)
pydantic-settings>=2.3.0   # settings from env vars

# Utilities
pyarrow>=16.1.0             # GeoParquet I/O
loguru>=0.7.2               # better logging

# Testing
pytest>=8.2.2
pytest-asyncio>=0.23.7
httpx>=0.27.0               # test client for FastAPI


# Install:  pip install -r requirements.txt

